requirements.txt
streamlit
pandas
plotly
python-dotenv
oracledb
google-generativeai